requirements.txt
tkintermapview
osmnx
networkx
Pillow
